move custom hook into component folder
